+++ Src/BlueDotBrigade.Weevil.Gui/Filter/FilterResultsView.xaml.cs
@@ -6,6 +6,7 @@
 	using System.Windows.Controls;
 	using BlueDotBrigade.Weevil.Data;
 	using BlueDotBrigade.Weevil.Diagnostics;
+	using BlueDotBrigade.Weevil.Gui.Threading;
 
 	/// <summary>
 	/// Interaction logic for FilterResultsView.xaml
@@ -26,7 +27,8 @@
 				}
 			};
 
-			this.DataContext = new FilterResultsViewModel(Application.Current.MainWindow, Application.Current.Dispatcher);
+			var uiDispatcher = new UiDispatcher(Application.Current.Dispatcher);
+			this.DataContext = new FilterResultsViewModel(Application.Current.MainWindow, uiDispatcher);
 
 			InitializeComponent();
 
@@ -87,8 +89,17 @@
 
 		private void ListView_OnSelectionChanged(object sender, SelectionChangedEventArgs e)
 		{
-			this.ViewModel.Select(e.AddedItems.Cast<IRecord>().ToList());
-			this.ViewModel.UnSelect(e.RemovedItems.Cast<IRecord>().ToList());
+			var added = e.AddedItems.Cast<IRecord>().ToList();
+			if (added.Count > 0)
+			{
+				this.ViewModel.Select(added);
+			}
+
+			var removed = e.RemovedItems.Cast<IRecord>().ToList();
+			if (removed.Count > 0)
+			{
+				this.ViewModel.UnSelect(removed);
+			}
 		}
 	}
 }