+++ Src/BlueDotBrigade.Weevil.Gui/Properties/AssemblyInfo.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Reflection;
 using System.Resources;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Windows;
 
@@ -53,6 +54,11 @@ using System.Windows;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("2.5.2.*")]
-[assembly: AssemblyFileVersion("2.5.2.0")]
+[assembly: AssemblyVersion("2.6.0.*")]
+[assembly: AssemblyFileVersion("2.6.0.0")]
 [assembly: NeutralResourcesLanguage("en")]
+
+#if DEBUG
+[assembly: InternalsVisibleTo("BlueDotBrigade.Weevil.Gui-UiTests")]
+[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")] // required by Moq
+#endif