+++ Tst/BlueDotBrigade.Weevil.Core-IntegrationTests/Properties/AssemblyInfo.cs
@@ -14,5 +14,5 @@ using System.Runtime.InteropServices;
 
 [assembly: Guid("d52fd49c-67d4-4d57-9b8e-806514b69e51")]
 
-[assembly: AssemblyVersion("2.5.2.*")]
-[assembly: AssemblyFileVersion("2.5.2.0")]
+[assembly: AssemblyVersion("2.6.0.*")]
+[assembly: AssemblyFileVersion("2.6.0.0")]